// apps/web/src/lib/api/generated/types/jobOrderEntityQuality.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type JobOrderEntityQuality = typeof JobOrderEntityQuality[keyof typeof JobOrderEntityQuality];


export const JobOrderEntityQuality = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;