// apps/web/src/lib/api/generated/types/jobOrderEntityStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type JobOrderEntityStatus = typeof JobOrderEntityStatus[keyof typeof JobOrderEntityStatus];


export const JobOrderEntityStatus = {
  ACTIVE: 'ACTIVE',
  PLACED: 'PLACED',
  CLOSED: 'CLOSED',
  ON_HOLD: 'ON_HOLD',
} as const;